Extract out minimal classes
