+++ src/main/java/Storage.java
@@ -1,0 +1,53 @@
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+import java.util.Scanner;
+
+public class Storage {
+    private static final String FILE_PATH = "data/sunoo.txt";
+
+    public static void loadTasks() throws IOException {
+        ensureFileExists();
+        Scanner s = new Scanner(new File(FILE_PATH));
+        while (s.hasNextLine()) {
+            String taskText = s.nextLine();
+            String[] taskParts = taskText.split(" \\| ");
+            boolean isDone;
+            switch (taskParts[0]) {
+            case "T":
+                isDone = taskParts[1].equals("1");
+                TaskList.addTask(new ToDo(isDone, taskParts[2]));
+                break;
+            case "D":
+                isDone = taskParts[1].equals("1");
+                TaskList.addTask(new Deadline(isDone, taskParts[2],
+                        LocalDateTime.parse(taskParts[3], DateTimeFormatter.ISO_LOCAL_DATE_TIME)));
+                break;
+            case "E":
+                isDone = taskParts[1].equals("1");
+                TaskList.addTask(new Event(isDone, taskParts[2],
+                        LocalDateTime.parse(taskParts[3], DateTimeFormatter.ISO_LOCAL_DATE_TIME),
+                        LocalDateTime.parse(taskParts[4], DateTimeFormatter.ISO_LOCAL_DATE_TIME)));
+            }
+        }
+    }
+
+    public static void updateTaskListInTxt() throws IOException {
+        FileWriter fw = new FileWriter(FILE_PATH);
+        for (Task task : TaskList.getTasks()) {
+            fw.write(task.getTxtRepresentation());
+            fw.write("\n");
+        }
+        fw.close();
+    }
+
+    private static void ensureFileExists() throws IOException {
+        File file = new File(FILE_PATH);
+        file.getParentFile().mkdirs();
+        file.createNewFile();
+    }
+}
